% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_slopes.R
\name{calc_slope}
\alias{calc_slope}
\title{Calculate slopes of models metrics=f(time)}
\usage{
calc_slope(datDGO, y_trans = "identity")
}
\arguments{
\item{datDGO}{}
}
\value{
a data table with slopes
}
\description{
Calculate slopes of models metrics=f(time)
}
