% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Get and format raw data from datapath}
\usage{
coverage(data_density, metric)
}
\arguments{
\item{data}{data aggregated by time period: comprises data_aggregated and data_density}
}
\value{
a tibble containing variables x, metric (name) and value, and possibly xcat
}
\description{
Get and format raw data from datapath
}
\examples{
data=readr::read_csv("data-raw/data_ganga.csv")
data=aggregate_data(data, time_acc="year")
coverage(data$data_density, "ACw_mean")
}
