% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_spacetime.R
\name{cut_space}
\alias{cut_space}
\title{Categorize based on space or time variable}
\usage{
cut_space(x, breaks, prefix = "R")
}
\arguments{
\item{type}{can be either "numeric" or "date". Defaults to "numeric"}
}
\value{
a categorical variable
}
\description{
Categorize based on space or time variable
}
\examples{
data_raw=readr::read_csv("data-raw/data_ganga.csv",show_col_types=FALSE)
data_summary=aggregate_data(data_raw, time_acc="year", space_acc=1)
cut_space(data_summary$data_aggregated$ID,breaks="50;100")
}
