% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_spacetime.R
\name{cut_spacetime}
\alias{cut_spacetime}
\title{Categorize based on space or time variable}
\usage{
cut_spacetime(x, breaks, type = "numeric", prefix = "R")
}
\arguments{
\item{type}{can be either "numeric" or "date". Defaults to "numeric"}
}
\value{
a categorical variable
}
\description{
Categorize based on space or time variable
}
