% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_spacetime.R
\name{cut_time}
\alias{cut_time}
\title{Categorize based on space or time variable}
\usage{
cut_time(x, breaks, prefix = "P")
}
\arguments{
\item{breaks}{}
}
\value{
a categorical variable
}
\description{
Categorize based on space or time variable
}
\examples{
data_raw=readr::read_csv("data-raw/data_ganga.csv",show_col_types=FALSE)
data_summary=aggregate_data(data_raw, time_acc="year", space_acc=1)
cut_time(data_summary$data_aggregated$DATE,breaks="2000;2005")
}
