% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get and format raw data from datapath}
\usage{
get_data(datapath, vartime = "DATE", varspace = "Unnamed: 0", vary)
}
\value{
a tibble containing variables x, metric (name) and value, and possibly xcat
}
\description{
Get and format raw data from datapath
}
\examples{
data=get_data("data/Lhassa_6_04_2024.csv",vary="VEGETATION_POLYGONS_p40")
data=get_data("data/Sabarmati_19_02_2024.csv",vary="VEGETATION_POLYGONS_p40")
}
