% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_default}
\alias{get_default}
\title{If corresponding input does not exist, return default value}
\usage{
get_default(input, name, value)
}
\arguments{
\item{input}{list of input elements}

\item{name}{name of input}

\item{value}{default value}
}
\value{
the default value if input does not exist or the input value if it does.
}
\description{
If corresponding input does not exist, return default value
}
\examples{
data=readr::read_csv("data-raw/data_Lhassa.csv")
dat=aggregate_data(data, time_acc="year")
}
