% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metric.R
\name{get_metric}
\alias{get_metric}
\title{Plots a metric through var_x}
\usage{
get_metric(
  data,
  metric = "ACw_mean",
  var_space = "ID",
  var_time = "DATE",
  breaks_space = "",
  breaks_time = ""
)
}
\arguments{
\item{data}{the data table}

\item{metric}{the name of the metric}

\item{var_space}{the variable to display as X variable}

\item{var_time}{the variable to display as X variable}

\item{breaks_space}{the x values where to place cuts (defaults to NULL)}

\item{breaks_time}{the x values where to place cuts (defaults to NULL)}
}
\value{
a tibble containing variables x, metric (name) and value, and possibly xcat
}
\description{
Plots a metric through var_x
}
\examples{
data=readr::read_csv("data/data_ganga.csv")
dat=get_metric(data)
get_metric(data,breaks_space="50;100")
get_metric(data,breaks_time="2000-07-01")
data=readr::read_csv("data/data_Lhassa.csv")
get_metric(data, metric="VEGETATION_POLYGONS_p40",breaks_time="2010-07-01")
}
