% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_slopes.R
\name{get_slopes}
\alias{get_slopes}
\title{Calculate slopes of models metrics=f(DATE)}
\usage{
get_slopes(data, y_trans = "identity")
}
\value{
a data table with slopes
}
\description{
Calculate slopes of models metrics=f(DATE)
}
\examples{

data=readr::read_csv("data/data_ganga.csv")
dat=get_metric(data,metric="ACw_mean")
get_slopes(dat)
}
