% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_cats.R
\name{label_cats}
\alias{label_cats}
\title{Labels categories of x for displaying in plot}
\usage{
label_cats(dat, x)
}
\arguments{
\item{x}{the name of the variable used for X axis}

\item{data}{data}
}
\value{
a vector with the labels for the categories of x
}
\description{
Labels categories of x for displaying in plot
}
\examples{
dat=get_metric(data,breaks_time="1998;2007",breaks_space="60")
label_cats(dat,"x_space")
}
