% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineplot_metric.R
\name{lineplot_metric}
\alias{lineplot_metric}
\title{Plots a metric through var_x with lineplots}
\usage{
lineplot_metric(
  dat,
  x = "x_space",
  color = "x_space_cat",
  scale_y = "identity",
  add_regression = FALSE
)
}
\arguments{
\item{x}{the name of the variable used for X axis}

\item{color}{the name of the variable used to color}

\item{scale_y}{the scale for Y variable (defaults to "identity")}

\item{add_regression}{whether to add a regression line}

\item{data}{the data table}

\item{add_means}{whether to add means (as red dots)}
}
\value{
lineplots showing the values of the metric through X
}
\description{
Plots a metric through var_x with lineplots
}
\examples{
data=readr::read_csv("data/data_ganga.csv")
dat=get_metric(data,breaks_time="2002-07-01",breaks_space="60")
lineplot_metric(dat)
lineplot_metric(dat,scale_y="log10")
lineplot_metric(dat,scale_y="log10", color="space_cat")
}
