% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_slopes.R
\name{plot_slopes}
\alias{plot_slopes}
\title{Plot slopes of models metrics=f(DATE)}
\usage{
plot_slopes(dat_slopes, seg = FALSE)
}
\arguments{
\item{dat_slopes}{}
}
\value{
a data table with slopes
}
\description{
Plot slopes of models metrics=f(DATE)
}
\examples{
data=readr::read_csv("data/data_ganga.csv")
dat=get_metric(data,metric="ACw_mean")
dat_slopes=get_slopes(dat,y_trans="log10")
plot_slopes(dat_slopes,seg=TRUE)
}
